Identify Load Balancers by K8S service UID
This commit adds a hcloud-ccm/service-uid label with the value of the
Kubernetes service UID to each Load Balancer created by the CCM. This
allows to uniquely identify Load Balancers managed by the CCM.

All Load Balancer related operations now use the Kubernetes service UID
to obtain the affected Load Balancer from the Hetzner Cloud Backend.
Thereby, renaming of Load Balancers can now be implemented properly as
the CCM does not need to know the old name of the Load Balancer anymore.

To allow the import of existing load balancers the CCM falls back to the
load-balancer.hetzner.cloud/name if no hcloud-ccm/service-uid label is
set. In this case it adds the label to the Load Balancer during
reconciliation.

Closes #128