Automate release process
The new release process uses goreleaser to automatically create and
publish releases. Among other things this allows us to make sure that
all version numbers are updated automatically based on the git tag we
set.

For every release after this commit the deployment example files are
published as a GitHub release. The links to the corresponding files in
the repository do not work anymore.