Add Master is attached to configured Network check on controller boot.
We had several reports/issues from customers who configured there CCM wrong. They copied there configuration and changed nothing (It uses the same API token and the configured network as the other cluster). However, the servers of the new cluster were in a different project and/or attached to a different network. Because of this, two CCM instances started to fight against each other, because the CCM always tries to reach the wanted state in the network, so it detaches everything it does not know about and then attaches his "known" routes.

With this MR we now check if the master is attached to the configured network by using a "not user controllable" source of truth - our metadata service, that is only reachable through the node public interface. This check will be performed on boot and if the node is not attached to the correct network, it will fail booting up with a message that gives this hint. I also added a new ENV to disable this check for cluster specific reasons (sample: Rancher clusters where the controlplan might be out of the cluster itself). We decided to check this "per default", so there might be issues for customers who have currently a somehow broken cluster but didn't noticed it.

Signed-off-by: Lukas Kämmerling <lukas.kaemmerling@hetzner-cloud.de>