chore: initial basic helm chart
chart templates created by running the following:

 $ tail -n+2 deploy/ccm.yaml | kubectl slice -t '{{.kind | lower}}.yaml' -o chart/templates/

generate-deployment-yaml.sh is updated to then do the previous thing in
reverse. That is, it runs `helm template chart` and the contents go in
deploy/ccm.yaml

As a result of this change, deploy/ccm.yaml.tmpl is no longer needed.

In follow-up commits the dev and network flavors will be generated from
the new chart, and the corresponding .tmpl files will go away.