refactor: unnecessary API call in instance reconciliation
Calls to InstanceV2.InstanceMetadata() caused 2 API calls, one
to retrieve the server and the second to check if the configured
network exists.

The second call is not necessary because we already know all
networks that the server is attached to exist. By just comparing
the nameOrID locally we can achieve the same result.