fix: self-reported version not correct
The release process was changed after v1.9.1 to use `ldflags` with `-X` to
dynamically set the version in the binary.

This silently failed because with `-X` you can only change the value of
_variables_, not _constants_ but our `providerVersion` was a constant
beforehand. You also need to provide the full import path, and not the
relative path from where you build the binary.

I also changed the default value to "unknown" to better reflect the value if
the binary was not properly built.