ci: build helm repo index (#391)
We previously relied on the helm-charts repo to re-build the index. But
this has problems:

- The gh-pages pipeline is not properly started because events created
by the default GITHUB_TOKEN do not trigger pipelines (in the same repo).
- Adding the chart and re-indexing happens in two commits, slightly
obfuscating the chain of events
- When running `helm repo index` it updates the created timestamp of all
existing entries. While I have not seen any issues yet, there might be
some clients that rely on this field.