chore: multiple improvements to the release process (#394)
### chore: fix release missing deployment files

The path for the deployment files was changed in
https://github.com/hetznercloud/hcloud-cloud-controller-manager/pull/375
but not updated here. This cause the v1.14.1 release to be missing the
release artifacts until I manually uploaded them.

### chore: no draft releases

In the current release process we need to "publish" the release twice,
once to create the tag and trigger `goreleaser`, and then a second time
because goreleaser turned the release back into a draft.

We can just skip the drafting step and let goreleaser publish directly.

### chore: simplify chart version update

Simplify the command used to update the Chart version and add some
comments.