feat: new dev/test environment (#414)
This PR simplifies and unifies the process of bringing up a hcloud k8s
environment that is suitable for development/testing purposes. It builds
off the work originally started in hetznercloud/csi-driver#226

The central script is `hack/dev-up.sh`. The script requires a
`HCLOUD_TOKEN`, and will build a 1+ k8s cluster using `k3sup` + `hcloud`
CLI tools. It's typically quite fast: a cold execution of the script
should take less <1min before a fully operational cluster is ready for
use.

The `dev-down.sh` script will delete all resources created by
`dev-up.sh`.

One of the primary objectives in this work was to simplify the e2e
testing process. Julian and I decided to remove the matrix of kubeadm
clusters, and also remove flannel as an option for CNI. Instead, our
test suite now runs on the latest 3 stable releases of k3s. Further, the
option to create a cluster with/without hcloud networking support has
been removed. Created clusters *always* have a network and an install of
Cilium configured for direct routing mode (traffic bound for pods on
other nodes is handed to hcloud network fabric).

Skaffold is plumbed into the e2e test pipelines on GitHub. This way, the
process of building and deploying a test build of hccm to a cluster is
unified for development and test.

Once this work has landed and stabilized here, we expect to package it
up a bit further and use it in csi-driver and anywhere else we operate
Kubernetes integrations that need automated testing.

---------

Co-authored-by: Julian Tölle <julian.toelle@hetzner-cloud.de>