fix(routes): let k/cloud-provider decide which routes to delete
We implemented a route cleanup in #238 that deletes all routes in the
network that can not be mapped to the private IP of a node. This causes
issues for users who create their own routes, e.g. to an alias IP.

We do not need to implement this ourselves, because
`kubernetes/cloud-provider` already implemented all the checks for us,
with more consideration about which routes actually need to be deleted.

In particular, the ccm should only delete routes that are within the
configured Cluster CIDR range.