test: use configurable domain for managed certificate (#451)
While migrating this project from TTS -> TPS we switched the Hetzner
account. The new account does not have the domain `hc-certs.de`
configured. While the test case still succeeds (does not check for cert
status), we get lots of emails for the failed certificate issuance.

This commit makes the domain used in the test configurable and skips
the test in case no domain is set.