ci: fix failure on main when pushing latest image
Since reworking the docker build process in #417 we do not have a
":latest" image to push. The CI pipeline in main has failed since then.

Getting a "latest" image is not that easy with the current goreleaser
setup. We use the goreleaser docker_manifests feature to build
multi-arch images. This stage is not executed in "--snapshot" builds.