feat(helm): allow to set labels and annotations for podMonitor (#471)
This PR adds the ability to set labels and annotations to the
podMonitor. Setting the label is usually necessary for the podMonitor to
be picked up by the Prometheus Operator
(https://github.com/prometheus-operator/prometheus-operator/issues/3636#issuecomment-796688902).
Setting an annotation helps to defer the deployment of the custom
resource if e.g. deploying with ArgoCD
(https://argo-cd.readthedocs.io/en/stable/user-guide/sync-options/#skip-dry-run-for-new-custom-resources-types).