ci: fix helm chart publishing
The helm chart was not published since 1.15.0. The script was not even
called, because a change between 1.15.0-rc.0 & 1.15.0 caused the archive
step to not be executed anymore.

As we never did anything with the archive anyway, I have removed it now.

Instead we ensure the "execute once" requirement for the chart publish
by depending on the checksum output.