ci: fix flake in dev/e2e env where subnet is not available
The e2e tests fail regularly with error message:

    hcloud: network 3334615 has no free IP available or is in a different network zone (invalid_input)

As far as I can tell, this happens because the network does not yet have
an available subnet when the server is created. With this change, we
will wait until the network has the subnet (by grepping for the subnet
cidr) before creating the server.