:bug: Fix InstanceExists for baremetal servers
Fixing the InstanceExists function for bare metal servers. The endpoint
did not check whether the node name matches and therefore always returns
that a certain bare metal instance exists, as these servers are treated
as pet and do not get removed from the system, as is the case for HCloud servers.

Furthermore, this commit changes the caching behavior to never call the
API for individual servers. This was not needed for InstanceV1
interface, as it worked with node names more often than providerIDs
which led to list calls of robot API anyway. This is not the case in
InstanceV2, where almost all the time get calls are used. These calls
can be cached by an occasional list call as we do now.