chore: skaffold was not using the built image
We changed the image reference in a recent PR. With this, skaffold was
unable to match it against the image referenced in the Helm chart, so it
always deployed the tagged version of the image.

This caused all e2e tests to not use the actual changes made in those
PRs.