ci: fix flaky e2e tests by colocating registry & hccm
In our dev/test environment we deploy a container registry to the
cluster and push our image there to avoid a dependency on external
registries.

To pull from the registry, containerd is instructed to use the service
IP. This IP is resolved to the Pod IP in kube-proxy. The CNI is then
responsible for making a connection to the Pod IP.

In our case we have the CNI configured to use the Cloud Routes for
accessing the Pod IPs from other nodes. This is only being setup by
HCCM, which is a problem if we need the routes for pulling the HCCM
container image.

This fixes the circular dependency by running HCCM on the same node as
the registry. That way no external routes are required and the traffic
stays on the same node.

This is not a problem on clusters where HCCM is pulled from an external
registry or where the routes controller is not used.