fix: populate ingress private ip when disable-private-ingress is false (#715)
Closes #553

Refactor getting the ingresses for a load-balancer for both get
load-balancer and ensure load-balancer, into using a shared function.
This ensures we are consistent between the 2 methods.


### Upgrading

The load balancer ingress IPs are now populated with the private IPs,
unless the `load-balancer.hetzner.cloud/disable-private-ingress`
annotation is set to `true`. Please make sure that you configured the
annotation according to your needs, for example if you are using
`external-dns`.