fix: wrong version logged on startup (#729)
The k/cloud-provider library is logging a version on startup. This
version is expected to be set by Kubernetes build scripts, but we do not
use this.

Right now it looks like this:

```
controllermanager.go:169] Version: v0.0.0-master+$Format:%H$
```

With the fix, this will report the same version as we use for the
user-agent.