feat(metrics): add metrics from cloud-provider library
This includes metrics about internal operations from
`k8s.io/cloud-provider` like the workqueue depth and requests to the
Kubernetes API.

This metrics were already exposed on `:8233/metrics` but this was not
documented or scraped.

This commit now uses the same registry for our metrics and the
kubernetes libraries, and also exposes them on both ports for backwards
compatibility.

Co-authored-by: Lukas Metzner <lukas.metzner@hetzner-cloud.de>