fix(metrics): do not emit kubernetes_build_info metric (#849)
Because `kubernetes_build_info` reports the version of the HCCM and not
Kubernetes, we are removing it from the metrics registry. This happens
because the intended use of `k8s.io/compeonent-base/metrics` expects us
to follow the Kubernetes release cycle and versioning. As the metric is
initialized in the `k8s.io/component-base/metrics/prometheus/version`
package, we can just remove the import.

This was discovered through alerts being raised by the default alert
rule `KubeVersionMismatch` of the `kube-prometheus-stack`.

---------

Co-authored-by: Julian Tölle <julian.toelle@hetzner-cloud.de>