feat(robot): forward InternalIPs by default on Robot nodes (#865)
Whenever a user configures the `--node-ip` flag, we forward the
configured IPs. We validate if the IP already exists as an ExternalIP
and does not collide with the configured address family. This improves
the Robot support for HCCM, as users can configure private IPs for Robot
nodes.

Forwarding only happens during the initialization phase of HCCM and
should therefore not break existing clusters.

This feature can be disabled by setting the environment variable
`ROBOT_FORWARD_INTERNAL_IPS` to `false`.

---------
Co-authored-by: Alexander Block <ablock84@gmail.com>
Co-authored-by: Julian Tölle <julian.toelle@hetzner-cloud.de>