ci: fix pre-commit golang version
As there was no `depName` defined with the `golang-version` `datasource`, our custom renovate regex manager did not detect this as a dependency. We can also use the git tags from the official repository as [`golang-version`](https://docs.renovatebot.com/modules/datasource/golang-version/) is using this as a source AFAIK.

Question is, if we really need this check at all. Pre-commit will use the newest golang version by default, which should be sufficient as `golangci-lint` will use the go version specified in the `go.mod`.