fix(load-balancer): don't print entire node object when adding robot target (#904)
When adding new targets to a loadbalancer,
hcloud-cloud-controller-manager usually prints something along the lines

    add target op="hcops/LoadBalancerOps.ReconcileHCLBTargets" service="<service_name>" targetName="<node_name>"

When adding robot nodes, it instead prints 

  add target op="hcops/LoadBalancerOps.ReconcileHCLBTargets" service="<service_name>"
targetName="<the entire serialized node object>"

which is quite obviously wrong. This PR fixes so that it is only the name that is
printed.

I also changed to log message to communicate that this it is a robot
node that is being added.