ci: ensure goreleaser custom publish runs only once
All artifacts must have a non empty ID, otherwise the `ids: [""]` filter will
include artifacts with an empty ID (strings in go are initialized to
"").

Add an ID to all artifacts to prevent running in the problem described
abovei and filter all IDs using a unused ID ("none") in the helm chart
publisher.

In addition, do skip publishing the helm chart if the chart file already
exists in the chart repository.